requirements.txt
ftputil==5.0.4
PyYAML==6.0
